+++ internal/jimm/bootstrap/mocks/juju_client_store.go
