+++ internal/jobtracker/jobtracker_test.go
@@ -144,11 +144,6 @@ func (s *jobTrackerSuite) TestStopJob(c *qt.C) {
 	s.pollJob(testCtx, id, c, dbmodel.StatusFailed)
 }
 
-//go:generate mockgen -destination=./mocks/bootstrapjobstore.go -package=mocks . BootstrapJobStore
-//go:generate mockgen -destination=./mocks/bootstrapjobjujumanager.go -package=mocks . BootstrapJobJujuManager
-//go:generate mockgen -destination=./mocks/bootstrapjobclistore.go -package=mocks . BootstrapJobBinaryStore
-//go:generate mockgen -destination=./mocks/bootstrapexecutor.go -package=mocks . BootstrapExecutor
-//go:generate mockgen -destination=./mocks/juju_client_store.go -package=mocks github.com/juju/juju/jujuclient ClientStore
 func TestJobTrackerSuite(t *testing.T) {
 	qtsuite.Run(qt.New(t), &jobTrackerSuite{})
 }
